feat: share logic between web and mobile
