--- packages/shared/getStartAt.ts
@@ -1,9 +1,0 @@
-export const getStartAt = (initialRemainingTime: number, duration: number) => {
-  if (duration === 0) {
-    return 0
-  }
-
-  return typeof initialRemainingTime === 'number'
-    ? duration - initialRemainingTime
-    : 0
-}
