+++ packages/shared/index.ts
@@ -1,45 +1,3 @@
-export const getPathProps = (
-  size: number,
-  strokeWidth: number,
-  rotation: 'clockwise' | 'counterclockwise'
-) => {
-  const halfSize = size / 2
-  const halfStrokeWith = strokeWidth / 2
-  const arcRadius = halfSize - halfStrokeWith
-  const arcDiameter = 2 * arcRadius
-  const rotationIndicator = rotation === 'clockwise' ? '1,0' : '0,1'
-
-  const pathLength = 2 * Math.PI * arcRadius
-  const path = `m ${halfSize},${halfStrokeWith}
-          a ${arcRadius},${arcRadius} 0 ${rotationIndicator} 0,${arcDiameter}
-          a ${arcRadius},${arcRadius} 0 ${rotationIndicator} 0,-${arcDiameter}`
-
-  return { path, pathLength }
-}
-
-export const getStartAt = (duration: number, initialRemainingTime?: number) => {
-  if (duration === 0) {
-    return 0
-  }
-
-  return typeof initialRemainingTime === 'number'
-    ? duration - initialRemainingTime
-    : 0
-}
-
-export const getWrapperStyle = (size: number): React.CSSProperties => ({
-  position: 'relative',
-  width: size,
-  height: size,
-})
-
-export const timeStyle: React.CSSProperties = {
-  display: 'flex',
-  justifyContent: 'center',
-  alignItems: 'center',
-  position: 'absolute',
-  left: 0,
-  top: 0,
-  width: '100%',
-  height: '100%',
-}
+export { useCountdown } from './useCountdown'
+export { getWrapperStyle, timeStyle } from './utils'
+export type { Props } from './types'
